+++ src/gui/widgets/shape_style/fill_style_widget.hpp
@@ -10,8 +10,9 @@ class FillStyleWidget : public ColorSelector
 public:
     FillStyleWidget(QWidget* parent = nullptr);
 
-    void set_shape(model::Fill* target, int gradient_stop = 0);
-    model::Fill* shape() const;
+    void set_targets(const std::vector<model::Fill*>& targets);
+    model::Fill* current() const;
+    void set_current(model::Fill* current);
 
     void set_gradient_stop(model::Styler* styler, int index);
 
@@ -19,6 +20,9 @@ private:
     void update_from_target();
 
     void set_color(const QColor& color, bool commit);
+
+    void before_set_target();
+    void after_set_target();
 
 private slots:
     void set_target_color(const QColor& color);
@@ -30,9 +34,10 @@ private slots:
     void clear_target_color();
 
 private:
-    model::Fill* target = nullptr;
+    std::vector<model::Styler*> targets;
+    model::Styler* current_target = nullptr;
     utils::PseudoMutex updating;
-    int stop = 0;
+    int stop = -1;
 };
 
 } // namespace glaxnimate::gui