+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Editing:
     * Added the ability to select the first bezier node from the node's context menu
+    * Setting fill / stroke / gradient updates all selected objects, not just the last selected
 * UI:
     * Nicer display of color values in tree views